refactor(server): split channel authn middleware
Closes #188
